+++ doc/api/search/classes_0.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['commandablehttpclient',['CommandableHttpClient',['../class_pip_services_1_1_rpc_1_1_clients_1_1_commandable_http_client.html',1,'PipServices3::Rpc::Clients']]],
-  ['commandablehttpservice',['CommandableHttpService',['../class_pip_services_1_1_rpc_1_1_services_1_1_commandable_http_service.html',1,'PipServices3::Rpc::Services']]]
-];