+++ doc/api/search/classes_3.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['iregisterable',['IRegisterable',['../interface_pip_services_1_1_rpc_1_1_services_1_1_i_registerable.html',1,'PipServices3::Rpc::Services']]]
-];