+++ doc/api/search/classes_5.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['statusrestservice',['StatusRestService',['../class_pip_services_1_1_rpc_1_1_services_1_1_status_rest_service.html',1,'PipServices3::Rpc::Services']]]
-];