+++ doc/api/search/functions_3.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['executeasync',['ExecuteAsync',['../class_pip_services_1_1_rpc_1_1_clients_1_1_rest_client.html#a243075b49ef1e9fffdb6d08e16647077',1,'PipServices3.Rpc.Clients.RestClient.ExecuteAsync(string correlationId, HttpMethod method, string route)'],['../class_pip_services_1_1_rpc_1_1_clients_1_1_rest_client.html#a80629949da4d6b563b2eb37881027fcc',1,'PipServices3.Rpc.Clients.RestClient.ExecuteAsync(string correlationId, HttpMethod method, string route, object requestEntity)']]],
-  ['executeasync_3c_20t_20_3e',['ExecuteAsync&lt; T &gt;',['../class_pip_services_1_1_rpc_1_1_clients_1_1_rest_client.html#ad44e87a1a75ae3043d6be0da6ef29796',1,'PipServices3.Rpc.Clients.RestClient.ExecuteAsync&lt; T &gt;(string correlationId, HttpMethod method, string route)'],['../class_pip_services_1_1_rpc_1_1_clients_1_1_rest_client.html#ae25c6c775db4060ddbe74dca08950e9a',1,'PipServices3.Rpc.Clients.RestClient.ExecuteAsync&lt; T &gt;(string correlationId, HttpMethod method, string route, object requestEntity)']]]
-];