+++ docs/search/classes_1.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['defaultrpcfactory',['DefaultRpcFactory',['../class_pip_services_1_1_rpc_1_1_build_1_1_default_rpc_factory.html',1,'PipServices3::Rpc::Build']]],
-  ['directclient',['DirectClient',['../class_pip_services_1_1_rpc_1_1_clients_1_1_direct_client.html',1,'PipServices3::Rpc::Clients']]]
+  ['basicauthorizer',['BasicAuthorizer',['../class_pip_services3_1_1_rpc_1_1_auth_1_1_basic_authorizer.html',1,'PipServices3::Rpc::Auth']]]
 ];