+++ docs/search/classes_4.js
@@ -1,5 +1,8 @@
 var searchData=
 [
-  ['restclient',['RestClient',['../class_pip_services_1_1_rpc_1_1_clients_1_1_rest_client.html',1,'PipServices3::Rpc::Clients']]],
-  ['restservice',['RestService',['../class_pip_services_1_1_rpc_1_1_services_1_1_rest_service.html',1,'PipServices3::Rpc::Services']]]
+  ['heartbeatoperations',['HeartbeatOperations',['../class_pip_services3_1_1_rpc_1_1_services_1_1_heartbeat_operations.html',1,'PipServices3::Rpc::Services']]],
+  ['heartbeatrestservice',['HeartbeatRestService',['../class_pip_services3_1_1_rpc_1_1_services_1_1_heartbeat_rest_service.html',1,'PipServices3::Rpc::Services']]],
+  ['httpconnectionresolver',['HttpConnectionResolver',['../class_pip_services3_1_1_rpc_1_1_connect_1_1_http_connection_resolver.html',1,'PipServices3::Rpc::Connect']]],
+  ['httpendpoint',['HttpEndpoint',['../class_pip_services3_1_1_rpc_1_1_services_1_1_http_endpoint.html',1,'PipServices3::Rpc::Services']]],
+  ['httpresponsesender',['HttpResponseSender',['../class_pip_services3_1_1_rpc_1_1_services_1_1_http_response_sender.html',1,'PipServices3::Rpc::Services']]]
 ];