+++ docs/search/functions_2.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['defaultrpcfactory',['DefaultRpcFactory',['../class_pip_services_1_1_rpc_1_1_build_1_1_default_rpc_factory.html#af2a819acc07d52728f085dae732e6841',1,'PipServices3::Rpc::Build::DefaultRpcFactory']]],
-  ['directclient',['DirectClient',['../class_pip_services_1_1_rpc_1_1_clients_1_1_direct_client.html#a68effecb901e674c0712c79e7ac1ee4a',1,'PipServices3::Rpc::Clients::DirectClient']]]
+  ['defaultrpcfactory',['DefaultRpcFactory',['../class_pip_services3_1_1_rpc_1_1_build_1_1_default_rpc_factory.html#a450d0ee44181f2da1936fa8a26d906cc',1,'PipServices3::Rpc::Build::DefaultRpcFactory']]],
+  ['directclient',['DirectClient',['../class_pip_services3_1_1_rpc_1_1_clients_1_1_direct_client.html#a9e8e2fbee0373c8a70f7322c3a482948',1,'PipServices3::Rpc::Clients::DirectClient']]]
 ];