+++ docs/search/namespaces_0.js
@@ -1,9 +1,10 @@
 var searchData=
 [
-  ['build',['Build',['../namespace_pip_services_1_1_rpc_1_1_build.html',1,'PipServices3::Rpc']]],
-  ['clients',['Clients',['../namespace_pip_services_1_1_rpc_1_1_clients.html',1,'PipServices3::Rpc']]],
-  ['connect',['Connect',['../namespace_pip_services_1_1_rpc_1_1_connect.html',1,'PipServices3::Rpc']]],
-  ['pipservices',['PipServices3',['../namespace_pip_services.html',1,'']]],
-  ['rpc',['Rpc',['../namespace_pip_services_1_1_rpc.html',1,'PipServices3']]],
-  ['services',['Services',['../namespace_pip_services_1_1_rpc_1_1_services.html',1,'PipServices3::Rpc']]]
+  ['auth',['Auth',['../namespace_pip_services3_1_1_rpc_1_1_auth.html',1,'PipServices3::Rpc']]],
+  ['build',['Build',['../namespace_pip_services3_1_1_rpc_1_1_build.html',1,'PipServices3::Rpc']]],
+  ['clients',['Clients',['../namespace_pip_services3_1_1_rpc_1_1_clients.html',1,'PipServices3::Rpc']]],
+  ['connect',['Connect',['../namespace_pip_services3_1_1_rpc_1_1_connect.html',1,'PipServices3::Rpc']]],
+  ['pipservices3',['PipServices3',['../namespace_pip_services3.html',1,'']]],
+  ['rpc',['Rpc',['../namespace_pip_services3_1_1_rpc.html',1,'PipServices3']]],
+  ['services',['Services',['../namespace_pip_services3_1_1_rpc_1_1_services.html',1,'PipServices3::Rpc']]]
 ];