+++ src/Services/RestService.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using PipServices3.Commons.Config;
+using PipServices3.Commons.Data;
 using PipServices3.Commons.Errors;
 using PipServices3.Commons.Refer;
 using PipServices3.Commons.Run;
@@ -28,7 +29,7 @@ namespace PipServices3.Rpc.Services
     /// - controller:            override for Controller dependency
     /// 
     /// connection(s):
-    /// - discovery_key:         (optional) a key to retrieve the connection from <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-components-dotnet/master/doc/api/interface_pip_services_1_1_components_1_1_connect_1_1_i_discovery.html">IDiscovery</a>
+    /// - discovery_key:         (optional) a key to retrieve the connection from <a href="https://pip-services3-dotnet.github.io/pip-services3-components-dotnet/interface_pip_services_1_1_components_1_1_connect_1_1_i_discovery.html">IDiscovery</a>
     /// - protocol:              connection protocol: http or https
     /// - host:                  host name or IP address
     /// - port:                  port number
@@ -36,10 +37,10 @@ namespace PipServices3.Rpc.Services
     /// 
     /// ### References ###
     /// 
-    /// - *:logger:*:*:1.0         (optional) <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-components-dotnet/master/doc/api/interface_pip_services_1_1_components_1_1_log_1_1_i_logger.html">ILogger</a> components to pass log messages
-    /// - *:counters:*:*:1.0         (optional) <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-components-dotnet/master/doc/api/interface_pip_services_1_1_components_1_1_count_1_1_i_counters.html">ICounters</a> components to pass collected measurements
-    /// - *:discovery:*:*:1.0        (optional) <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-components-dotnet/master/doc/api/interface_pip_services_1_1_components_1_1_connect_1_1_i_discovery.html">IDiscovery</a> services to resolve connection
-    /// - *:endpoint:http:*:1.0          (optional) <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-rpc-dotnet/master/doc/api/class_pip_services_1_1_rpc_1_1_services_1_1_http_endpoint.html">HttpEndpoint</a> reference
+    /// - *:logger:*:*:1.0         (optional) <a href="https://pip-services3-dotnet.github.io/pip-services3-components-dotnet/interface_pip_services_1_1_components_1_1_log_1_1_i_logger.html">ILogger</a> components to pass log messages
+    /// - *:counters:*:*:1.0         (optional) <a href="https://pip-services3-dotnet.github.io/pip-services3-components-dotnet/interface_pip_services_1_1_components_1_1_count_1_1_i_counters.html">ICounters</a> components to pass collected measurements
+    /// - *:discovery:*:*:1.0        (optional) <a href="https://pip-services3-dotnet.github.io/pip-services3-components-dotnet/interface_pip_services_1_1_components_1_1_connect_1_1_i_discovery.html">IDiscovery</a> services to resolve connection
+    /// - *:endpoint:http:*:1.0          (optional) <a href="https://pip-services3-dotnet.github.io/pip-services3-rpc-dotnet/class_pip_services_1_1_rpc_1_1_services_1_1_http_endpoint.html">HttpEndpoint</a> reference
     /// </summary>
     /// <example>
     /// <code>
@@ -333,7 +334,32 @@ namespace PipServices3.Rpc.Services
         {
             return HttpResponseSender.SendDeletedResultAsync(response, result);
         }
-        
+
+        protected string GetCorrelationId(HttpRequest request)
+        {
+            return HttpRequestHelper.GetCorrelationId(request);
+        }
+
+        protected FilterParams GetFilterParams(HttpRequest request)
+        {
+            return HttpRequestHelper.GetFilterParams(request);
+        }
+
+        protected PagingParams GetPagingParams(HttpRequest request)
+        {
+            return HttpRequestHelper.GetPagingParams(request);
+        }
+
+        protected SortParams GetSortParams(HttpRequest request)
+        {
+            return HttpRequestHelper.GetSortParams(request);
+        }
+
+        protected RestOperationParameters GetParameters(HttpRequest request)
+        {
+            return HttpRequestHelper.GetParameters(request);
+        }
+
         private string AppendBaseRoute(string route) {
             if (!string.IsNullOrEmpty(_baseRoute)) {
                 var baseRoute = _baseRoute;
@@ -383,6 +409,15 @@ namespace PipServices3.Rpc.Services
             _endpoint.RegisterInterceptor(route, action);
         }
 
+        protected virtual void RegisterSwaggerRoute(string method, string route,
+             Func<HttpRequest, HttpResponse, RouteData, Task> action)
+        {
+            if (_endpoint == null) return;
+
+            route = AppendBaseRoute(route);
+            _endpoint.RegisterSwaggerRoute(method, route, action);
+        }
+
         public virtual void Register()
         { }
 
@@ -413,7 +448,7 @@ namespace PipServices3.Rpc.Services
             {
                 var responseContent = content;
 
-                RegisterRoute(HttpMethods.Get, _swaggerRoute, async (request, response, routeData) =>
+                RegisterSwaggerRoute(HttpMethods.Get, _swaggerRoute, async (request, response, routeData) =>
                 {
                     response.ContentType = "application/json";
                     response.StatusCode = (int)HttpStatusCode.OK;